Added status resolver, to replace explicit access to the hits attribute
The status resolver has been introduced to remove any explicit access to the character's hits attribute. This made code inherently coupled tightly to the USR combat/game system by relying on the presence of that attribute to check a universal rpg character state (dead or alive.) The resolver allows for one to check this state without relying on a specific game system implementation.